[#1] Refactoring
