+++ internal/backend/neofs/util.go
@@ -2,7 +2,6 @@ package neofs
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"strconv"
@@ -10,14 +9,10 @@ import (
 
 	"github.com/nspcc-dev/neo-go/cli/flags"
 	"github.com/nspcc-dev/neo-go/pkg/wallet"
-	"github.com/nspcc-dev/neofs-api-go/pkg/acl"
-	"github.com/nspcc-dev/neofs-api-go/pkg/client"
 	"github.com/nspcc-dev/neofs-api-go/pkg/container"
 	cid "github.com/nspcc-dev/neofs-api-go/pkg/container/id"
-	"github.com/nspcc-dev/neofs-api-go/pkg/netmap"
 	"github.com/nspcc-dev/neofs-api-go/pkg/object"
 	"github.com/nspcc-dev/neofs-api-go/pkg/owner"
-	"github.com/nspcc-dev/neofs-sdk-go/policy"
 	"github.com/nspcc-dev/neofs-sdk-go/pool"
 )
 
@@ -29,8 +24,6 @@ type BuffCloser struct {
 func (bc *BuffCloser) Close() error {
 	return nil
 }
-
-var ErrNotFound = errors.New("not found")
 
 func createPool(ctx context.Context, cfg Config) (pool.Pool, error) {
 	acc, err := getAccount(cfg)
@@ -74,13 +67,12 @@ func getAccount(cfg Config) (*wallet.Account, error) {
 	return acc, nil
 }
 
-func getContainerID(ctx context.Context, client pool.Pool, containerName, placementPolicy string) (*cid.ID, error) {
-	cnrID, err := findContainerID(ctx, client, containerName)
-	if err != nil && errors.Is(err, ErrNotFound) {
-		return createContainer(ctx, client, containerName, placementPolicy)
+func getContainerID(ctx context.Context, client pool.Pool, container string) (*cid.ID, error) {
+	cnrID := cid.New()
+	if err := cnrID.Parse(container); err != nil {
+		return findContainerID(ctx, client, container)
 	}
-
-	return cnrID, err
+	return cnrID, nil
 }
 
 func findContainerID(ctx context.Context, client pool.Pool, containerName string) (*cid.ID, error) {
@@ -102,52 +94,7 @@ func findContainerID(ctx context.Context, client pool.Pool, containerName string
 		}
 	}
 
-	return nil, ErrNotFound
-}
-
-func createContainer(ctx context.Context, client pool.Pool, containerName, placementPolicy string) (*cid.ID, error) {
-	pp, err := policy.Parse(placementPolicy)
-	if err != nil {
-		return nil, err
-	}
-
-	cnr := container.New(
-		container.WithPolicy((*netmap.PlacementPolicy)(pp)),
-		container.WithCustomBasicACL(acl.PrivateBasicRule),
-		container.WithAttribute(container.AttributeName, containerName),
-		container.WithAttribute(container.AttributeTimestamp, strconv.FormatInt(time.Now().Unix(), 10)))
-	cnr.SetOwnerID(client.OwnerID())
-
-	containerID, err := client.PutContainer(ctx, cnr)
-	if err != nil {
-		return nil, err
-	}
-
-	err = waitPresence(ctx, client, containerID)
-	return containerID, err
-}
-
-func waitPresence(ctx context.Context, cli client.Container, cnrID *cid.ID) error {
-	wctx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-	ticker := time.NewTimer(5 * time.Second)
-	defer ticker.Stop()
-	wdone := wctx.Done()
-	done := ctx.Done()
-	for {
-		select {
-		case <-done:
-			return ctx.Err()
-		case <-wdone:
-			return wctx.Err()
-		case <-ticker.C:
-			_, err := cli.GetContainer(ctx, cnrID)
-			if err == nil {
-				return nil
-			}
-			ticker.Reset(5 * time.Second)
-		}
-	}
+	return nil, fmt.Errorf("container '%s' not found", containerName)
 }
 
 func formRawObject(own *owner.ID, cnrID *cid.ID, name string, header map[string]string) *object.RawObject {
